@@ -1,6 +1,7 @@
 import logging
 import os
 import re
+from configparser import ConfigParser
 from subprocess import SubprocessError
 from time import sleep, time
 
@@ -19,9 +20,12 @@ class AlembicMigrationHelper:
         alembic_directory_path: str = "/app/alembic",
         alembic_config_file_path: str = "alembic.ini",
         database_url: str,
+        alembic_section_name: str,
         additional_parameters: str = "",
         allow_migration_for_empty_database: bool = False,
         configure: bool = True,
+        # skip_db_checks is used for testing purposes only
+        skip_db_checks: bool = False,
     ):
         if not database_url:
             raise ValueError("database_url not set, not upgrading database.")
@@ -31,30 +35,42 @@ class AlembicMigrationHelper:
         self.additional_parameters = additional_parameters
         self.alembic_directory_path = alembic_directory_path
         self.alembic_config_file_path = alembic_config_file_path
+        self.alembic_section_name = alembic_section_name
+        self.skip_db_checks = skip_db_checks
 
         # Chartreuse will upgrade a PG managed/configured by postgres-operator
         self.is_patroni_postgresql: bool = "CHARTREUSE_PATRONI_POSTGRESQL" in os.environ
-        if self.is_patroni_postgresql:
+        if self.is_patroni_postgresql and not skip_db_checks:
             self.additional_parameters += " -x patroni_postgresql=yes"
             self._wait_postgres_is_configured()
 
         if configure:
             self._configure()
-
-        self.is_migration_needed = self._check_migration_needed()
+        # skip_db_checks is used for testing purposes only
+        if not skip_db_checks:
+            self.is_migration_needed = self._check_migration_needed()
+        else:
+            self.is_migration_needed = False
 
     def _configure(self) -> None:
-        with open(f"{self.alembic_directory_path}/{self.alembic_config_file_path}") as f:
-            content = f.read()
-            content_new = re.sub(
-                "(sqlalchemy.url.*=.*){1}", r"sqlalchemy.url=%s" % self.database_url, content, flags=re.M
-            )
-        if content != content_new:
-            with open(f"{self.alembic_directory_path}/{self.alembic_config_file_path}", "w") as f:
-                f.write(content_new)
-            logger.info("alembic.ini was configured.")
-        else:
-            raise SubprocessError("configuration of alembic.ini has failed (alembic.ini is unchanged)")
+        config_path = f"{self.alembic_directory_path}/{self.alembic_config_file_path}"
+
+        # Read the configuration file
+        config = ConfigParser()
+        config.read(config_path)
+
+        # Multi-database configuration: update specific section
+        if not config.has_section(self.alembic_section_name):
+            raise ValueError(f"Section '{self.alembic_section_name}' not found in {self.alembic_config_file_path}")
+
+        # Update the sqlalchemy.url in the specific section
+        config.set(self.alembic_section_name, "sqlalchemy.url", self.database_url)
+
+        # Write the updated configuration back to the file
+        with open(config_path, "w") as f:
+            config.write(f)
+
+        logger.info("alembic.ini was configured for section %s.", self.alembic_section_name)
 
     def _wait_postgres_is_configured(self) -> None:
         """
@@ -91,7 +107,7 @@ class AlembicMigrationHelper:
                 return
             except Exception as e:
                 # TODO: Learn about exceptions that should be caught here, otherwise we'll wait for nothing
-                logger.info(f"Caught: {e}")
+                logger.info("Caught: %s", e)
                 logger.info(
                     "Waiting for the postgres-operator to create the user wiremind_owner_user"
                     " (that alembic and I use) and to set default privileges..."
@@ -107,7 +123,7 @@ class AlembicMigrationHelper:
 
     def is_postgres_empty(self) -> bool:
         table_list = self._get_table_list()
-        logger.info(f"Tables in the database: {table_list}")
+        logger.info("Tables in the database: %s", table_list)
         # Don't count "alembic" table
         table_name = "alembic_version"
         if table_name in table_list:
