@@ -11,4 +11,7 @@ def configure_chartreuse_mock(mocker: MockerFixture, is_migration_needed: bool =
         create=True,
     )
     mocker.patch("chartreuse.chartreuse_upgrade.Chartreuse.upgrade")
-    mocker.patch("wiremind_kubernetes.kubernetes_helper._get_namespace_from_kube", return_value="foo")
+    mocker.patch(
+        "wiremind_kubernetes.kubernetes_helper._get_namespace_from_kube",
+        return_value="foo",
+    )
