@@ -2,49 +2,57 @@ import pytest
 from pytest_mock.plugin import MockerFixture
 
 import chartreuse.chartreuse_upgrade
-from chartreuse import get_version
 
 from ..conftest import configure_os_environ_mock
 from .conftest import configure_chartreuse_mock
 
 
-def test_chartreuse_upgrade_detected_migration_enabled_stop_pods(mocker: MockerFixture) -> None:
+def test_chartreuse_upgrade_detected_migration_enabled_stop_pods(
+    mocker: MockerFixture,
+) -> None:
     """
     Test that chartreuse_upgrades stop pods in case of detected migration.
     """
     configure_chartreuse_mock(mocker=mocker, is_migration_needed=True)
     mocked_stop_pods = mocker.patch("wiremind_kubernetes.KubernetesDeploymentManager.stop_pods")
     mocker.patch("wiremind_kubernetes.KubernetesDeploymentManager.start_pods")
-    configure_os_environ_mock(mocker=mocker, additional_environment={"HELM_CHART_VERSION": get_version()})
+    mocker.patch("chartreuse.chartreuse_upgrade.get_version", return_value="5.0.0")
+    configure_os_environ_mock(mocker=mocker, additional_environment={"HELM_CHART_VERSION": "5.0.0"})
 
     chartreuse.chartreuse_upgrade.main()
     mocked_stop_pods.assert_called()
 
 
-def test_chartreuse_upgrade_detected_migration_disabled_stop_pods(mocker: MockerFixture) -> None:
+def test_chartreuse_upgrade_detected_migration_disabled_stop_pods(
+    mocker: MockerFixture,
+) -> None:
     """
     Test that chartreuse_upgrades does not stop pods in case of detected migration but we disallow stop-pods.
     """
     configure_chartreuse_mock(mocker=mocker, is_migration_needed=True)
     mocked_stop_pods = mocker.patch("wiremind_kubernetes.KubernetesDeploymentManager.stop_pods")
     mocker.patch("wiremind_kubernetes.KubernetesDeploymentManager.start_pods")
+    mocker.patch("chartreuse.chartreuse_upgrade.get_version", return_value="5.0.0")
     configure_os_environ_mock(
         mocker=mocker,
-        additional_environment=dict(CHARTREUSE_ENABLE_STOP_PODS="", HELM_CHART_VERSION=get_version()),
+        additional_environment={"CHARTREUSE_ENABLE_STOP_PODS": "", "HELM_CHART_VERSION": "5.0.0"},
     )
 
     chartreuse.chartreuse_upgrade.main()
     mocked_stop_pods.assert_not_called()
 
 
-def test_chartreuse_upgrade_no_migration_disabled_stop_pods(mocker: MockerFixture) -> None:
+def test_chartreuse_upgrade_no_migration_disabled_stop_pods(
+    mocker: MockerFixture,
+) -> None:
     """
     Test that chartreuse_upgrades does NOT stop pods in case of migration not needed.
     """
     configure_chartreuse_mock(mocker=mocker, is_migration_needed=False)
     mocked_stop_pods = mocker.patch("wiremind_kubernetes.KubernetesDeploymentManager.stop_pods")
     mocker.patch("wiremind_kubernetes.KubernetesDeploymentManager.start_pods")
-    configure_os_environ_mock(mocker=mocker, additional_environment={"HELM_CHART_VERSION": get_version()})
+    mocker.patch("chartreuse.chartreuse_upgrade.get_version", return_value="5.0.0")
+    configure_os_environ_mock(mocker=mocker, additional_environment={"HELM_CHART_VERSION": "5.0.0"})
 
     chartreuse.chartreuse_upgrade.main()
     mocked_stop_pods.assert_not_called()
@@ -72,7 +80,10 @@ def test_chartreuse_upgrade_no_migration_disabled_stop_pods(mocker: MockerFixtur
     ],
 )
 def test_chartreuse_upgrade_compatibility_check(
-    mocker: MockerFixture, helm_chart_version: str | None, package_version: str, should_raise: bool
+    mocker: MockerFixture,
+    helm_chart_version: str | None,
+    package_version: str,
+    should_raise: bool,
 ) -> None:
     """
     Test that chartreuse_upgrade deals as expected with compatibility with the package version and the
